Unification of the names of the IO plugins
- They are all called xxxx_io_plugin now.
